Relocate Cometbft Initialization (#10)
* moved cometbft setup from run cmd to init cmd

* fmt.Println formating fix
